# Conflicts:
#	core/block/service.go
#	core/indexer/doc.go
#	core/indexer/indexer.go
#	core/indexer/indexer_test.go

+++ core/block/editor.go
@@ -29,6 +29,8 @@ import (
 	"github.com/gogo/protobuf/types"
 )
 
+var ErrOptionUsedByOtherObjects = fmt.Errorf("option is used by other objects")
+
 func (s *service) MarkArchived(id string, archived bool) (err error) {
 	return s.Do(id, func(b smartblock.SmartBlock) error {
 		return b.SetDetails(nil, []*pb.RpcBlockSetDetailsDetail{
@@ -200,6 +202,30 @@ func (s *service) ConvertChildrenToPages(req pb.RpcBlockListConvertChildrenToPag
 	}
 
 	return linkIds, err
+}
+
+func (s *service) UpdateBlockContent(ctx *state.Context, req pb.RpcBlockUpdateContentRequest) (err error) {
+	err = s.DoBasic(req.ContextId, func(b basic.Basic) error {
+		var found bool
+		err = b.Update(ctx, func(b simple.Block) error {
+			found = true
+			expectedType := fmt.Sprintf("%T", b.Model().GetContent())
+			gotType := fmt.Sprintf("%T", req.GetBlock().Content)
+			if gotType != expectedType {
+				return fmt.Errorf("block content should have %s type, got %s instead", expectedType, gotType)
+			}
+			b.Model().Content = req.GetBlock().Content
+			return nil
+		}, req.BlockId)
+		if err != nil {
+			return err
+		} else if !found {
+			return smartblock.ErrSimpleBlockNotFound
+		}
+
+		return nil
+	})
+	return
 }
 
 func (s *service) ReplaceBlock(ctx *state.Context, req pb.RpcBlockReplaceRequest) (newId string, err error) {
@@ -729,13 +755,32 @@ func (s *service) UpdateExtraRelationOption(ctx *state.Context, req pb.RpcObject
 }
 
 func (s *service) DeleteExtraRelationOption(ctx *state.Context, req pb.RpcObjectRelationOptionDeleteRequest) error {
-	return s.Do(req.ContextId, func(b smartblock.SmartBlock) error {
-		err := b.DeleteExtraRelationOption(ctx, req.RelationKey, req.OptionId, true)
-		if err != nil {
-			return err
+	objIds, err := s.anytype.ObjectStore().AggregateObjectIdsForOptionAndRelation(req.RelationKey, req.OptionId)
+	if err != nil {
+		return err
+	}
+
+	if !req.ConfirmRemoveAllValuesInRecords {
+		for _, objId := range objIds {
+			if objId != req.ContextId {
+				return ErrOptionUsedByOtherObjects
+			}
 		}
-		return nil
-	})
+	} else {
+		for _, objId := range objIds {
+			err = s.Do(objId, func(b smartblock.SmartBlock) error {
+				err := b.DeleteExtraRelationOption(ctx, req.RelationKey, req.OptionId, true)
+				if err != nil {
+					return err
+				}
+				return nil
+			})
+			if err != nil && err != smartblock.ErrRelationOptionNotFound {
+				return err
+			}
+		}
+	}
+	return nil
 }
 
 func (s *service) SetObjectTypes(ctx *state.Context, objectId string, objectTypes []string) (err error) {
